pull out gnome extension interfaces
